load-testing/requirements.txt
blinker==1.8.2
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.3.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
locust==2.30.0
MarkupSafe==2.1.5
msgpack==1.0.8
psutil==6.0.0
python-dotenv==1.0.1
pyzmq==26.1.0
requests==2.32.3
urllib3==2.2.2
Werkzeug==3.0.3
zope.event==5.0
zope.interface==6.4.post2
